add global func to handle ui updates
