+++ lib/guards/auth-guard.dart
@@ -50,6 +50,7 @@ Future<String?> unprotectedRouteRedirect(
       userModel, userPreferencesModel, categoryModel, tagModel);
   var redirectRoute = redirect ?? "/";
 
+  // TODO: find a better way to handle this
   if (tokensValid) {
     GroupSelectUIShellBuilder.setupBottomNav(context);
     appBarModel.setAppBarData(titleText: "Groups");
