+++ lib/shared/widgets/screen_wrapper.dart
@@ -7,6 +7,7 @@ import 'package:receipt_wrangler_mobile/groups/nav/receipt_ui_shell_builder.dart
 import 'package:receipt_wrangler_mobile/models/app_bar_model.dart';
 import 'package:receipt_wrangler_mobile/models/group_model.dart';
 import 'package:receipt_wrangler_mobile/utils/group.dart';
+import 'package:receipt_wrangler_mobile/utils/shell.dart';
 
 class ScreenWrapper extends StatefulWidget {
   const ScreenWrapper({
@@ -36,31 +37,7 @@ class _ScreenWrapper extends State<ScreenWrapper> {
     var router = GoRouter.of(context);
 
     router.routerDelegate.addListener(() {
-      var fullPath = router.routerDelegate.currentConfiguration.fullPath;
-
-      if (fullPath == "/groups") {
-        GroupSelectUIShellBuilder.setupBottomNav(context);
-        Provider.of<AppBarModel>(context, listen: false)
-            .setAppBarData(titleText: "Groups");
-      } else if (fullPath == "/groups/:groupId/dashboards") {
-        var groupId = getGroupByIdWithRouter(router) ?? "";
-        var group = Provider.of<GroupModel>(context, listen: false)
-            .getGroupById(groupId);
-
-        GroupDashboardUIShellBuilder.setupBottomNav(context);
-        Provider.of<AppBarModel>(context, listen: false).setAppBarData(
-            titleText: group!.name, leadingArrowRedirect: "/groups");
-      } else if (fullPath == "/groups/:groupId/receipts") {
-        var groupId = getGroupByIdWithRouter(router) ?? "";
-        var group = Provider.of<GroupModel>(context, listen: false)
-            .getGroupById(groupId);
-
-        Provider.of<AppBarModel>(context, listen: false).setAppBarData(
-            titleText: "${group!.name} Receipts",
-            leadingArrowRedirect: "/groups");
-      } else if (fullPath == "/receipts/:receiptId/view") {
-        ReceiptUIShellBuilder.setupBottomNav(context);
-      }
+      handleShellUIUpdate(context);
     });
   }
